Use re2 regular expression engine for bind path matching
To match bind paths in cases where the ':' character may be part of the
paths themselves and not just used as path separators, a complex regex
expression is used. This expression was borrowed from Docker CE, which is
written in Go. The issue is that Go uses a regex engine called re2[0],
whereas many languages (including Python used in this project) uses a more
more classic implementation. The classic engine, while faster in the
general case, suffers from an issue that causes it to go exponential on
some patterns, including the one used here for bind paths. This means
any binds longer than about ~30 charterers will effectively stall.

Use pyre2[1] to allow us to make use of the re2 engine.

[0] https://github.com/google/re2
[1] https://pypi.org/project/pyre2/